Changed class structure + work on multiple classes

+++ BabySurvivor/src/Projectile.cpp
@@ -1,6 +1,6 @@
 #include "Projectile.h"
 
-Projectile::Projectile(sf::Vector2f coords, int damage, bool isAlly) :
+Projectile::Projectile(sf::Vector2f coords, float damage, bool isAlly) :
 	coords{ coords },
 	damage{ damage },
 	isAlly{ isAlly } {}
